chore: move drawing to ryot_tiled
