+++ apps/ryot_compass/src/lmdb.rs
@@ -2,13 +2,7 @@ use crate::{ExportMap, LoadMap};
 use bevy::prelude::*;
 use heed::types::Bytes;
 use log::{debug, warn};
-use ryot::bevy_ryot::drawing::{update, CommandState};
-use ryot::bevy_ryot::lmdb::{
-    compact_map, read_area, reload_visible_area, LmdbCompactor, LmdbEnv,
-    LmdbPlugin as RyotLmdbPlugin,
-};
-use ryot::bevy_ryot::GameObjectBundle;
-use ryot::prelude::drawing::TileComponent;
+use ryot::prelude::LmdbPlugin as RyotLmdbPlugin;
 use ryot::prelude::*;
 use std::collections::HashMap;
 use std::sync::atomic::Ordering;
