+++ apps/ryot_compass/src/main.rs
@@ -4,7 +4,9 @@ use bevy::prelude::*;
 use bevy::window::PrimaryWindow;
 use bevy::winit::WinitWindows;
 use bevy_egui::EguiContexts;
-use ryot::prelude::*;
+use ryot::bevy_ryot::GamePlugin;
+#[cfg(feature = "lmdb")]
+use ryot_compass::lmdb::LmdbPlugin;
 use ryot_compass::*;
 use std::io::Cursor;
 use winit::window::Icon;
