+++ crates/ryot/src/bevy_ryot/game/mod.rs
@@ -1,4 +1,3 @@
-use crate::prelude::*;
 use bevy::prelude::*;
 use ryot_tiled::prelude::*;
 
@@ -11,25 +10,3 @@ impl Plugin for GamePlugin {
         app.add_event::<LoadObjects>().add_plugins(ElevationPlugin);
     }
 }
-
-#[derive(Bundle, Copy, Clone, Debug)]
-pub struct GameObjectBundle {
-    pub object_id: GameObjectId,
-    pub position: TilePosition,
-    pub elevation: Elevation,
-    pub layer: Layer,
-}
-
-impl GameObjectBundle {
-    pub fn new(object_id: GameObjectId, position: TilePosition, layer: Layer) -> Self {
-        Self {
-            object_id,
-            position,
-            layer,
-            elevation: default(),
-        }
-    }
-}
-
-#[derive(Event, Clone, Debug, Deref, DerefMut)]
-pub struct LoadObjects(pub Vec<GameObjectBundle>);
