+++ crates/ryot_tiled/src/map/lmdb/error.rs
