+++ crates/ryot_tiled/src/map/lmdb/serde.rs
